Support for DeePBSP v4 Extended nodes

+++ prboom2/src/r_defs.h
@@ -304,7 +304,9 @@ typedef struct
 typedef struct subsector_s
 {
   sector_t *sector;
-  unsigned short numlines, firstline;
+  // e6y: support for extended nodes
+  // make them an 'int' instead of a 'short'
+  int numlines, firstline;
 } subsector_t;
 
 
@@ -315,7 +317,8 @@ typedef struct
 {
   fixed_t  x,  y, dx, dy;        // Partition line.
   fixed_t bbox[2][4];            // Bounding box for each child.
-  unsigned short children[2];    // If NF_SUBSECTOR its a subsector.
+  //unsigned short children[2];    // If NF_SUBSECTOR its a subsector.
+  int children[2];    // If NF_SUBSECTOR its a subsector.
 } node_t;
 
 //
